backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary>=2.9.11
django-extensions==3.2.3
folium==0.15.1
geopy==2.4.1
openrouteservice==2.3.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt>=5.5.1